refactor(encrypt): 模块化设计加密逻辑
